fix(profileformgr): fixed bug where profile was not updating
BREAKING CHANGE: updated changes not appearing until after refresh
