+++ src/state/contexts/FormContext.js
@@ -1,9 +1,4 @@
-import React, { createContext, useContext, useState } from 'react';
-import { useHistory } from 'react-router-dom';
-
-import { UsersContext } from './UsersContext';
-import { GroomersContext } from './GroomersContext';
-import { APIContext } from './APIContext';
+import React, { createContext, useState } from 'react';
 
 export const FormContext = createContext({});
 
@@ -20,11 +15,6 @@ const FormProvider = ({ children }) => {
   //for delete modal
   const [showDelete, setShowDelete] = useState(false);
   // context state
-  const { userInfo, isRegistered } = useContext(UsersContext);
-  const { hours } = useContext(GroomersContext);
-  const { postUserInfo, putUserInfo } = useContext(APIContext);
-
-  const history = useHistory();
 
   // functions
   const toggleForm = () => {
@@ -33,35 +23,6 @@ const FormProvider = ({ children }) => {
 
   const onFailed = errorInfo => {
     setResultInfo({ message: 'Error: Please try again', type: 'error' });
-  };
-
-  const onFinish = (values, authState) => {
-    const hoursString = JSON.stringify(hours);
-    //add in user id and hours
-    const infoValues = {
-      user_id: userInfo.sub,
-      hours: hoursString,
-      ...values,
-    };
-
-    //checking isRegistered and calling the API to either create or update
-    //API calls are abstracted out into the API/index file as functions and called here
-    if (isRegistered === false) {
-      postUserInfo(
-        `${process.env.REACT_APP_API_URI}/groomers/`,
-        authState,
-        infoValues,
-        setResultInfo,
-        history
-      );
-    } else {
-      putUserInfo(
-        `${process.env.REACT_APP_API_URI}/groomers/${userInfo.sub}`,
-        authState,
-        infoValues,
-        setResultInfo
-      );
-    }
   };
 
   return (
@@ -84,7 +45,6 @@ const FormProvider = ({ children }) => {
         setNewValue,
         setShowDelModal,
         toggleForm,
-        onFinish,
         onFailed,
       }}
     >
